Command-line version.
Duplicated the code that creates the graph from hello.py into git-workflow.py.

Ideally, that code could be moved to a library, but I'm not familiar with heroku set-up and I don't want to break it.

I also added the gist that expands bash and git aliases, so the program can take output straight from the history command.

A prunned version of the requirements.txt is also provided.